requirements.txt
anyio==4.4.0
APScheduler==3.10.4
attrs==23.2.0
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
pycparser==2.22
PySocks==1.7.1
python-dotenv==1.0.1
python-telegram-bot==21.4
pytz==2024.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
nest-asyncio==1.5.8
playwright==1.20.0
fastapi==0.110.0
uvicorn==0.30.0
quart==0.19.5
undetected-chromedriver==3.5.5
selenium==4.18.1
